Refactor: make carousel item a component
